Updated result determination
