+++ Rock.php
@@ -2,6 +2,11 @@
 
 declare(strict_types=1);
 
+require_once 'SignInterface.php';
+require_once 'TieResult.php';
+require_once 'WinResult.php';
+require_once 'LoseResult.php';
+
 class Rock extends Result implements SignInterface
 {
     protected array $beateable = [
@@ -11,15 +16,13 @@ class Rock extends Result implements SignInterface
     public function beats(SignInterface $element): Result
     {
         if ($this instanceof $element) {
-            return new WinResult();
+            return new TieResult();
         }
 
         if (in_array(get_class($element), $this->beateable)) {
-            return new LoseResult();
+            return new WinResult();
         }
 
-        return new TieResult();
+        return new LoseResult();
     }
 }
-    
-
